refactor(engineer_worktree/tests): split into tests + tests_more (#1266)
tests.rs 458 -> 275; tests_more.rs 283 LOC. Helpers duplicated.
