+++ src/engineer_worktree/tests.rs
@@ -272,187 +272,3 @@ fn allocate_without_main_branch_returns_hard_error_and_leaves_no_dir() {
 // ---------------------------------------------------------------------------
 // Test 6 — startup sweep removes orphan dirs but leaves live worktrees alone
 // ---------------------------------------------------------------------------
-
-#[test]
-fn sweep_removes_orphan_dirs_and_preserves_live_worktrees() {
-    let parent_dir = tempdir().unwrap();
-    let state_dir = tempdir().unwrap();
-    let parent_repo = init_parent_repo(parent_dir.path());
-
-    let live = EngineerWorktree::allocate(&parent_repo, state_dir.path(), "goal-live")
-        .expect("allocate live");
-
-    let orphan = state_dir
-        .path()
-        .join("engineer-worktrees")
-        .join("goal-orphan-9999999999-deadbe");
-    fs::create_dir_all(&orphan).expect("create orphan dir");
-    fs::write(orphan.join("stale"), b"x").unwrap();
-
-    let report =
-        sweep_orphaned_worktrees(&parent_repo, state_dir.path()).expect("sweep must succeed");
-
-    assert!(
-        report.removed_orphan_dirs.iter().any(|p| p == &orphan),
-        "orphan {} must be reported as removed; got {:?}",
-        orphan.display(),
-        report.removed_orphan_dirs
-    );
-    assert!(!orphan.exists(), "orphan dir must be removed from disk");
-
-    assert!(live.path().exists(), "live worktree dir must remain");
-    assert!(
-        worktree_registered(&parent_repo, live.path()),
-        "live worktree registration must survive sweep"
-    );
-
-    live.cleanup().unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Test 7 — observation scope: parent-repo edits are invisible inside worktree
-//
-// This is the issue-#1197 root-cause test: a sibling/operator write to the
-// shared checkout MUST NOT show up inside an engineer's own worktree.
-// ---------------------------------------------------------------------------
-
-#[test]
-fn verification_scope_isolates_worktree_from_parent_repo_mutations() {
-    let parent_dir = tempdir().unwrap();
-    let state_dir = tempdir().unwrap();
-    let parent_repo = init_parent_repo(parent_dir.path());
-
-    let wt =
-        EngineerWorktree::allocate(&parent_repo, state_dir.path(), "goal-iso").expect("allocate");
-
-    let before: Vec<_> = fs::read_dir(wt.path())
-        .unwrap()
-        .filter_map(Result::ok)
-        .map(|e| e.file_name())
-        .collect();
-
-    fs::write(parent_repo.join("sibling-write.txt"), b"intruder").unwrap();
-
-    let after: Vec<_> = fs::read_dir(wt.path())
-        .unwrap()
-        .filter_map(Result::ok)
-        .map(|e| e.file_name())
-        .collect();
-
-    assert_eq!(
-        before, after,
-        "parent-repo mutation must NOT be visible from the engineer worktree; \
-         this is the root-cause fix for issue #1197"
-    );
-
-    wt.cleanup().unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Test 8 — goal_id validation (F1): rejects path traversal, ref injection,
-// hidden-file leading-dot, and oversized inputs at the boundary.
-// ---------------------------------------------------------------------------
-
-#[test]
-fn rejects_invalid_goal_id() {
-    let parent_dir = tempdir().unwrap();
-    let state_dir = tempdir().unwrap();
-    let parent_repo = init_parent_repo(parent_dir.path());
-
-    let cases: &[&str] = &[
-        "",          // empty
-        "../../etc", // path traversal
-        "..",        // parent dir
-        ".hidden",   // leading dot
-        "-rf",       // leading dash (argv injection)
-        "has space", // disallowed byte
-        "has/slash", // disallowed byte
-        "has\nnewl", // control char
-    ];
-    for bad in cases {
-        let err = EngineerWorktree::allocate(&parent_repo, state_dir.path(), bad)
-            .expect_err(&format!("goal_id {bad:?} must be rejected"));
-        assert!(
-            matches!(err, SimardError::ActionExecutionFailed { .. }),
-            "expected ActionExecutionFailed for {bad:?}, got {err:?}"
-        );
-    }
-
-    // 65-byte input must fail; 64-byte must succeed.
-    let too_long = "a".repeat(65);
-    let err = EngineerWorktree::allocate(&parent_repo, state_dir.path(), &too_long)
-        .expect_err("65-byte goal_id must be rejected");
-    assert!(
-        matches!(err, SimardError::ActionExecutionFailed { .. }),
-        "got {err:?}"
-    );
-
-    let max_ok = "a".repeat(64);
-    let wt = EngineerWorktree::allocate(&parent_repo, state_dir.path(), &max_ok)
-        .expect("64-byte goal_id must be accepted");
-    wt.cleanup().expect("cleanup max-len worktree");
-
-    // Confirm the worktrees root was NOT polluted by any of the rejected ids.
-    let worktrees_root = state_dir.path().join("engineer-worktrees");
-    if worktrees_root.exists() {
-        for entry in fs::read_dir(&worktrees_root).unwrap().flatten() {
-            let name = entry.file_name();
-            let name = name.to_string_lossy();
-            assert!(
-                name.starts_with(&max_ok),
-                "rejected goal_id leaked to disk as {name:?}"
-            );
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Test 9 — sweep skips symlinks (F2/F3).
-// A symlink planted under engineer-worktrees/ pointing at an unrelated dir
-// must NOT be classified as an orphan and must NOT have its target deleted.
-// ---------------------------------------------------------------------------
-
-#[cfg(unix)]
-#[test]
-fn sweep_skips_symlinks_and_preserves_targets() {
-    use std::os::unix::fs::symlink;
-
-    let parent_dir = tempdir().unwrap();
-    let state_dir = tempdir().unwrap();
-    let target_dir = tempdir().unwrap();
-    let parent_repo = init_parent_repo(parent_dir.path());
-
-    // Create the worktrees root and plant a symlink inside it pointing at
-    // a directory whose contents must survive the sweep.
-    let worktrees_root = state_dir.path().join("engineer-worktrees");
-    fs::create_dir_all(&worktrees_root).unwrap();
-    let canary = target_dir.path().join("canary");
-    fs::write(&canary, b"do-not-delete").unwrap();
-
-    let link = worktrees_root.join("evil-symlink");
-    symlink(target_dir.path(), &link).expect("plant symlink");
-
-    let report = sweep_orphaned_worktrees(&parent_repo, state_dir.path())
-        .expect("sweep must succeed even with symlink present");
-
-    assert!(
-        report.removed_orphan_dirs.is_empty(),
-        "symlink must not be reported as removed orphan; got {:?}",
-        report.removed_orphan_dirs
-    );
-    assert!(
-        canary.exists(),
-        "symlink target contents must survive sweep"
-    );
-    // Symlink itself should still be there (skipped, not deleted).
-    assert!(
-        fs::symlink_metadata(&link).is_ok(),
-        "symlink should be left in place for an operator to investigate"
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Test 10 — main_sha must be 40-hex (F7).
-// Already covered by the no-main test; add an explicit shape check via the
-// happy path: branch must point at the resolved 40-hex sha.
-// ---------------------------------------------------------------------------
